requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp_socks==0.10.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.6.3
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
bottle==0.13.2
bottle-websocket==0.2.9
Brotli==1.1.0
browser-cookie3==0.20.1
cachetools==4.2.2
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clr_loader==0.2.7.post0
colorama==0.4.6
cssselect2==0.7.0
curl_cffi==0.9.0
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.18
duckduckgo_search==7.4.4
EbookLib==0.18
et_xmlfile==2.0.0
fastapi==0.115.8
Flask==3.1.0
frozenlist==1.5.0
future==1.0.0
g4f==0.4.7.3
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.1
lz4==4.4.3
MarkupSafe==3.0.2
moviepy==1.0.3
mss==10.0.0
multidict==6.1.0
nest-asyncio==1.6.0
nodriver==0.39
numpy==2.2.3
odfpy==1.4.1
openpyxl==3.1.5
pillow==10.4.0
platformdirs==4.3.6
plyer==2.1.0
proglog==0.1.10
propcache==0.3.0
proxy-tools==0.1.0
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
PyPDF2==3.0.1
python-aiml==0.9.3
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.20
python-socks==2.7.1
python-telegram-bot==13.15
pythonnet==3.0.5
pytz==2025.1
pywebview==5.4
pywin32==308
requests==2.32.3
shadowcopy==0.0.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
srt==3.5.3
starlette==0.45.3
tinycss2==1.4.0
tornado==6.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
urllib3==2.3.0
uvicorn==0.34.0
vosk==0.3.45
webencodings==0.5.1
websockets==15.0
Werkzeug==3.1.3
whisper==1.1.10
WMI==1.5.1
wrapt==1.17.2
yarl==1.18.3
zope.event==5.0
zope.interface==7.2
